+++ docs/src/nav.ts
@@ -48,10 +48,17 @@ export const nav: NavGroup[] = [
         href: '/react/handbook/composition',
       },
       {
+        label: 'Customization',
+        href: '/react/handbook/customization',
+      },
+      {
+        label: 'TypeScript',
+        href: '/react/handbook/typescript',
+      },
+      {
         label: 'llms.txt',
         external: true,
         href: '/llms.txt',
-        isNew: true,
       },
     ],
   },
@@ -65,6 +72,11 @@ export const nav: NavGroup[] = [
       {
         label: 'Alert Dialog',
         href: '/react/components/alert-dialog',
+      },
+      {
+        label: 'Autocomplete',
+        href: '/react/components/autocomplete',
+        isNew: true,
       },
       {
         label: 'Avatar',
@@ -81,6 +93,11 @@ export const nav: NavGroup[] = [
       {
         label: 'Collapsible',
         href: '/react/components/collapsible',
+      },
+      {
+        label: 'Combobox',
+        href: '/react/components/combobox',
+        isNew: true,
       },
       {
         label: 'Context Menu',
@@ -149,7 +166,6 @@ export const nav: NavGroup[] = [
       {
         label: 'Select',
         href: '/react/components/select',
-        isNew: true,
       },
       {
         label: 'Separator',