+++ packages/react/src/alert-dialog/trigger/AlertDialogTrigger.test.tsx
@@ -43,7 +43,7 @@ describe('<AlertDialog.Trigger />', () => {
       expect(screen.queryByText('title text')).to.equal(null);
 
       await user.keyboard('[Tab]');
-      expect(document.activeElement).to.not.equal(trigger);
+      expect(document.activeElement).not.to.equal(trigger);
     });
 
     it('custom element', async () => {
@@ -68,7 +68,7 @@ describe('<AlertDialog.Trigger />', () => {
       expect(screen.queryByText('title text')).to.equal(null);
 
       await user.keyboard('[Tab]');
-      expect(document.activeElement).to.not.equal(trigger);
+      expect(document.activeElement).not.to.equal(trigger);
     });
   });
 });