+++ packages/react/src/field/control/FieldControl.tsx
@@ -153,7 +153,7 @@ export const FieldControl = React.forwardRef(function FieldControl(
       getInputValidationProps(),
       elementProps,
     ],
-    customStyleHookMapping: fieldValidityMapping,
+    stateAttributesMapping: fieldValidityMapping,
   });
 
   return element;