+++ packages/react/src/number-field/increment/NumberFieldIncrement.tsx
@@ -6,7 +6,7 @@ import { useButton } from '../../use-button';
 import { useNumberFieldRootContext } from '../root/NumberFieldRootContext';
 import { useNumberFieldButton } from '../root/useNumberFieldButton';
 import type { NumberFieldRoot } from '../root/NumberFieldRoot';
-import { styleHookMapping } from '../utils/styleHooks';
+import { stateAttributesMapping } from '../utils/stateAttributesMapping';
 
 /**
  * A stepper button that increases the field value when clicked.
@@ -48,6 +48,8 @@ export const NumberFieldIncrement = React.forwardRef(function NumberFieldIncreme
     stopAutoChange,
     value,
     valueRef,
+    lastChangedValueRef,
+    onValueCommitted,
   } = useNumberFieldRootContext();
 
   const disabled = disabledProp || contextDisabled;
@@ -74,6 +76,8 @@ export const NumberFieldIncrement = React.forwardRef(function NumberFieldIncreme
     intentionalTouchCheckTimeout,
     movesAfterTouchRef,
     locale,
+    lastChangedValueRef,
+    onValueCommitted,
   });
 
   const { getButtonProps, buttonRef } = useButton({
@@ -93,7 +97,7 @@ export const NumberFieldIncrement = React.forwardRef(function NumberFieldIncreme
     ref: [forwardedRef, buttonRef],
     state: buttonState,
     props: [props, elementProps, getButtonProps],
-    customStyleHookMapping: styleHookMapping,
+    stateAttributesMapping,
   });
 
   return element;