+++ packages/react/src/select/root/SelectRoot.tsx
@@ -5,12 +5,11 @@ import { useMergedRefs } from '@base-ui-components/utils/useMergedRefs';
 import { useSelectRoot } from './useSelectRoot';
 import { SelectRootContext, SelectFloatingContext } from './SelectRootContext';
 import { useFieldRootContext } from '../../field/root/FieldRootContext';
-import { serializeValue } from '../utils/serialize';
+import { serializeValue } from '../../utils/serializeValue';
 import {
   type BaseUIEventDetails,
   createBaseUIEventDetails,
 } from '../../utils/createBaseUIEventDetails';
-import { BaseUIChangeEventReason } from '../../utils/types';
 
 /**
  * Groups all parts of the select.
@@ -304,6 +303,15 @@ export namespace SelectRoot {
     unmount: () => void;
   }
 
-  export type ChangeEventReason = BaseUIChangeEventReason | 'window-resize';
+  export type ChangeEventReason =
+    | 'trigger-press'
+    | 'outside-press'
+    | 'escape-key'
+    | 'window-resize'
+    | 'item-press'
+    | 'focus-out'
+    | 'list-navigation'
+    | 'cancel-open'
+    | 'none';
   export type ChangeEventDetails = BaseUIEventDetails<ChangeEventReason>;
 }