+++ packages/react/src/slider/root/SliderRoot.test.tsx
@@ -110,8 +110,6 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
       expect(root).to.have.attribute('aria-labelledby', 'labelId');
 
       expect(slider).to.have.attribute('aria-valuenow', '30');
-      expect(slider).to.have.attribute('aria-valuemin', '0');
-      expect(slider).to.have.attribute('aria-valuemax', '100');
       expect(slider).to.have.attribute('aria-orientation', 'horizontal');
       expect(slider).to.have.attribute('aria-labelledby', 'labelId');
       expect(slider).to.have.attribute('step', '1');
@@ -144,7 +142,7 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
 
   describe.skipIf(isJSDOM || isWebKit)('rtl', () => {
     it('should handle RTL', async () => {
-      const handleValueChange = spy();
+      const handleValueChange = spy((newValue) => newValue);
       const { getByTestId } = await render(
         <div dir="rtl">
           <DirectionProvider direction="rtl">
@@ -169,8 +167,8 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
       );
 
       expect(handleValueChange.callCount).to.equal(2);
-      expect(handleValueChange.args[0][0]).to.equal(80);
-      expect(handleValueChange.args[1][0]).to.equal(78);
+      expect(handleValueChange.firstCall.returnValue).to.equal(80);
+      expect(handleValueChange.lastCall.returnValue).to.equal(78);
     });
   });
 
@@ -455,9 +453,8 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
   });
 
   describe('prop: max', () => {
-    it('should set the max and aria-valuemax on the input', async () => {
+    it('sets the max attribute on the input', async () => {
       await render(<TestSlider defaultValue={150} step={100} max={750} />);
-      expect(screen.getByRole('slider')).to.have.attribute('aria-valuemax', '750');
       expect(screen.getByRole('slider')).to.have.attribute('max', '750');
     });
 
@@ -524,9 +521,8 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
   });
 
   describe('prop: min', () => {
-    it('should set the min and aria-valuemin on the input', async () => {
-      await render(<TestSlider defaultValue={150} step={100} min={150} />);
-      expect(screen.getByRole('slider')).to.have.attribute('aria-valuemin', '150');
+    it('sets the min attribute on the input', async () => {
+      await render(<TestSlider defaultValue={150} step={100} min={150} max={200} />);
       expect(screen.getByRole('slider')).to.have.attribute('min', '150');
     });
 
@@ -633,7 +629,11 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
     it.skipIf(isJSDOM || isWebKit)('should support touch events', async () => {
       const handleValueChange = spy();
       const { getByTestId } = await render(
-        <TestRangeSlider defaultValue={[20, 30]} onValueChange={handleValueChange} />,
+        <TestRangeSlider
+          defaultValue={[20, 30]}
+          style={{ width: '100px' }}
+          onValueChange={handleValueChange}
+        />,
       );
       const sliderControl = getByTestId('control');
       stub(sliderControl, 'getBoundingClientRect').callsFake(getHorizontalSliderRect);
@@ -1020,7 +1020,13 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
         const handleValueChange = spy();
 
         await render(
-          <TestRangeSlider min={0} max={5} onValueChange={handleValueChange} value={value} />,
+          <TestRangeSlider
+            min={0}
+            max={5}
+            onValueChange={handleValueChange}
+            value={value}
+            style={{ width: '100px' }}
+          />,
         );
 
         const sliderControl = screen.getByTestId('control');
@@ -1799,7 +1805,7 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
     });
 
     it('should receive name prop from Field.Root', async () => {
-      const { container } = await render(
+      await render(
         <Field.Root name="field-slider">
           <Slider.Root>
             <Slider.Control>
@@ -1809,8 +1815,7 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
         </Field.Root>,
       );
 
-      const hiddenInput = container.querySelector('input[aria-hidden="true"]');
-      expect(hiddenInput).to.have.attribute('name', 'field-slider');
+      expect(screen.getByRole('slider')).to.have.attribute('name', 'field-slider');
     });
 
     it('[data-touched]', async () => {
@@ -1879,7 +1884,7 @@ describe.skipIf(typeof Touch === 'undefined')('<Slider.Root />', () => {
       expect(root).not.to.have.attribute('data-focused');
     });
 
-    describe('prop: validate', async () => {
+    describe('prop: validate', () => {
       it('runs on blur by default', async () => {
         await render(
           <Field.Root validate={() => 'error'}>